+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * added `True-Client-IP` header to receive the real visitor IP (Cloudflare)
 * added new bot list by Atmire (https://github.com/atmire/COUNTER-Robots/blob/master/COUNTER_Robots_list.json)
 * added script to update and clean up the User-Agent blacklist
+* added flagging bots based on how quickly page views are being sent
 * fixed grouping referrers with path in URL
 * updated User-Agent blacklist
 * updated referrer blacklist